requirements.txt
praw
requests
beautifulsoup4
python-dotenv
